Refactor Terra viewer orchestration
